# src/Web.Client/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/invoices/number/{number}": {
      "get": {
        "tags": [
          "Invoice"
        ],
        "summary": "Получить счет по его номеру",
        "operationId": "GetInvoiceNumber",
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invoices": {
      "get": {
        "tags": [
          "Invoice"
        ],
        "summary": "Получить все счета",
        "operationId": "GetAllInvoices",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "x-position": 1
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            },
            "x-position": 2
          },
          {
            "name": "filterExpression",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "filterValues",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "nullable": true,
              "items": {
                "type": "string",
                "nullable": true
              }
            },
            "x-position": 4
          }
        ],
        "responses": {
          "200": {
            "description": "Список счетов",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceModel"
                  }
                }
              }
            }
          },
          "204": {
            "description": "Список счетов"
          }
        }
      },
      "post": {
        "tags": [
          "Invoice"
        ],
        "summary": "Создать счет",
        "operationId": "ExposeInvoice",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceCreateRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "guid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/invoices/ship/{invoiceId}": {
      "put": {
        "tags": [
          "Invoice"
        ],
        "summary": "Закрытие счета, смена статуса и выставление даты отгрузки",
        "operationId": "Invoice_Ship",
        "parameters": [
          {
            "name": "invoiceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "shipDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "202": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/invoices/{id}": {
      "put": {
        "tags": [
          "Invoice"
        ],
        "summary": "Обновление счета",
        "operationId": "Invoice_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceUpdateRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Invoice"
        ],
        "summary": "Удалить счет по его id",
        "operationId": "Invoice_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/techcard/{id}": {
      "get": {
        "tags": [
          "CompositionTechCard"
        ],
        "summary": "Получение тех.карты по Id",
        "operationId": "GetCompositionTechCardById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechCard"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "CompositionTechCard"
        ],
        "summary": "Удаление тех.карты",
        "operationId": "DeleteTechCard",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/tech-card-step/{techCardId}": {
      "get": {
        "tags": [
          "TechCardStep"
        ],
        "summary": "Получение всех этапов по Id тех.карты",
        "operationId": "GetTechCardStepsByTechCardId",
        "parameters": [
          {
            "name": "techCardId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TechCardStep"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tech-card-step": {
      "post": {
        "tags": [
          "TechCardStep"
        ],
        "summary": "Добавление этапа в тех.карту",
        "operationId": "CreateTechCardStep",
        "requestBody": {
          "x-name": "model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TechCardStepRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "guid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tech-card-step/{id}": {
      "delete": {
        "tags": [
          "TechCardStep"
        ],
        "summary": "Удаление этапа",
        "operationId": "DeleteTechCardStep",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/tech-card-item/{techCardId}": {
      "get": {
        "tags": [
          "TechCardItem"
        ],
        "summary": "Получение состава тех.карты по TechCardId",
        "operationId": "GetAllTechCardItems",
        "parameters": [
          {
            "name": "techCardId",
            "in": "path",
            "required": true,
            "description": "id тех.карты",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TechCardItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tech-card-item": {
      "post": {
        "tags": [
          "TechCardItem"
        ],
        "summary": "Добавления компонента в тех.карту",
        "operationId": "CreateTechCardItem",
        "requestBody": {
          "x-name": "model",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TechCardItemRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "guid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tech-card-item/{id}": {
      "delete": {
        "tags": [
          "TechCardItem"
        ],
        "summary": "Удаление компонента из тех.карты",
        "operationId": "DeleteTechCardItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/storages": {
      "post": {
        "tags": [
          "Storage"
        ],
        "summary": "Создание нового склада",
        "operationId": "CreateStorage",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorageCreateRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "get": {
        "tags": [
          "Storage"
        ],
        "summary": "Получить список всех складов",
        "operationId": "GetAllStorages",
        "responses": {
          "200": {
            "description": "Список складов и общее количество",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Storage"
                  }
                }
              }
            }
          },
          "204": {
            "description": "Список складов и общее количество"
          }
        }
      }
    },
    "/api/v1/steps/{id}": {
      "get": {
        "tags": [
          "Step"
        ],
        "summary": "Получение этапа по Id",
        "operationId": "GetStepById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Step"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Step"
        ],
        "summary": "Удаление этапа",
        "operationId": "Step_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Invoice": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AggregateRoot"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "number": {
                "type": "string"
              },
              "contractorId": {
                "type": "string",
                "format": "guid"
              },
              "description": {
                "type": "string",
                "nullable": true
              },
              "paymentPercentage": {
                "type": "number",
                "format": "decimal"
              },
              "daysShipment": {
                "type": "integer",
                "format": "int32"
              },
              "status": {
                "$ref": "#/components/schemas/InvoiceStatus"
              },
              "dateShipment": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              }
            }
          }
        ]
      },
      "InvoiceStatus": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SmartEnumerationOfInvoiceStatus"
          },
          {
            "type": "object",
            "additionalProperties": false
          }
        ]
      },
      "SmartEnumerationOfInvoiceStatus": {
        "type": "object",
        "x-abstract": true,
        "additionalProperties": false,
        "properties": {
          "value": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "AggregateRoot": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DomainAuditableEntity"
          },
          {
            "type": "object",
            "x-abstract": true,
            "additionalProperties": false,
            "properties": {
              "domainEvents": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/IDomainEvent"
                }
              }
            }
          }
        ]
      },
      "IDomainEvent": {
        "type": "object",
        "x-abstract": true,
        "additionalProperties": false
      },
      "DomainAuditableEntity": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DomainEntity"
          },
          {
            "type": "object",
            "x-abstract": true,
            "additionalProperties": false,
            "properties": {
              "createdAt": {
                "type": "string",
                "format": "date-time"
              },
              "createdById": {
                "type": "string",
                "format": "guid"
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time"
              },
              "updatedById": {
                "type": "string",
                "format": "guid"
              },
              "deletedAt": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              },
              "deletedById": {
                "type": "string",
                "format": "guid",
                "nullable": true
              }
            }
          }
        ]
      },
      "DomainEntity": {
        "type": "object",
        "x-abstract": true,
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          }
        }
      },
      "InvoiceModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "number": {
            "type": "string"
          },
          "contractorId": {
            "type": "string",
            "format": "guid"
          },
          "contractorName": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "paymentPercentage": {
            "type": "number",
            "format": "decimal"
          },
          "daysShipment": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/InvoiceStatus"
          },
          "dateShipment": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "InvoiceCreateRequest": {
        "type": "object",
        "description": "Модель запроса на создание счета",
        "additionalProperties": false,
        "properties": {
          "number": {
            "type": "string"
          },
          "contragentId": {
            "type": "string",
            "format": "guid"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "daysShipment": {
            "type": "integer",
            "format": "int32"
          },
          "paymentPercentage": {
            "type": "number",
            "format": "decimal"
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceProduct"
            }
          }
        }
      },
      "InvoiceProduct": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DomainEntity"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "invoiceId": {
                "type": "string",
                "format": "guid"
              },
              "productId": {
                "type": "string",
                "format": "guid"
              },
              "quantity": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ]
      },
      "InvoiceUpdateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "number": {
            "type": "string"
          },
          "contractorId": {
            "type": "string",
            "format": "guid"
          },
          "issueStatus": {
            "$ref": "#/components/schemas/IssueStatus"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "daysShipment": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "paymentPercentage": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "dateShipment": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "IssueStatus": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SmartEnumerationOfIssueStatus"
          },
          {
            "type": "object",
            "additionalProperties": false
          }
        ]
      },
      "SmartEnumerationOfIssueStatus": {
        "type": "object",
        "x-abstract": true,
        "additionalProperties": false,
        "properties": {
          "value": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "UnitOfMeasure": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Piece"
        ],
        "enum": [
          1
        ]
      },
...
      "TechCard": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AggregateRoot"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "number": {
                "type": "string"
              },
              "description": {
                "type": "string",
                "nullable": true
              },
              "productId": {
                "type": "string",
                "format": "guid"
              }
            }
          }
        ]
      },
...
      "TechCardStep": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DomainEntity"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "techCardId": {
                "type": "string",
                "format": "guid"
              },
              "stepId": {
                "type": "string",
                "format": "guid"
              },
              "number": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ]
      },
      "TechCardStepRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "techCardId": {
            "type": "string",
            "format": "guid"
          },
          "stepId": {
            "type": "string",
            "format": "guid"
          },
          "number": {
            "type": "integer",
            "description": "Номер в очереди\n            ",
            "format": "int32"
          }
        }
      },
      "TechCardItem": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DomainEntity"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "techCardId": {
                "type": "string",
                "format": "guid"
              },
              "productId": {
                "type": "string",
                "format": "guid"
              },
              "quantity": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ]
      },
      "TechCardItemRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "techCardId": {
            "type": "string",
            "format": "guid"
          },
          "productId": {
            "type": "string",
            "description": "Ссылка на продукт, из которого состоит целевой продукт в тех.карте\n            ",
            "format": "guid"
          },
          "quantity": {
            "type": "integer",
            "description": "Необходимое количество, для создания единицы целевого продукта\n            ",
            "format": "int32"
          }
        }
      },
      "StorageCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "responsibleId": {
            "type": "string",
            "format": "guid"
          }
        }
      },
      "Storage": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DomainAuditableEntity"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "name": {
                "type": "string"
              },
              "description": {
                "type": "string",
                "nullable": true
              },
              "responsibleId": {
                "type": "string",
                "format": "guid"
              }
            }
          }
        ]
      },
      "Step": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DomainEntity"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "techCardId": {
                "type": "string",
                "format": "guid"
              },
              "operationId": {
                "type": "string",
                "format": "guid"
              },
              "description": {
                "type": "string",
                "nullable": true
              },
              "duration": {
                "type": "string",
                "format": "duration"
              },
              "cost": {
                "type": "number",
                "format": "decimal"
              },
              "unitOfMeasure": {
                "$ref": "#/components/schemas/UnitOfMeasure"
              }
            }
          }
        ]
      },
...
    }
  }
}